requirements.txt
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
